pyproject.toml
[tool.poetry]
name = "groqai-vectorsearch-by-streamlit"
version = "0.1.0"
description = "an AI-Chatbot with vector search using Groq AI API."
authors = ["Shogo Ogami <sg.tao.so@gmail.com>"]
license = "Apache-2.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.34.0"
groq = "^0.6.0"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.8.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
torch = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
